+++ src/middleware/index.ts
@@ -3,4 +3,4 @@ export { NonceTxMiddleware, isInvalidTxNonceError } from './nonce-tx-middleware'
 export { CachedNonceTxMiddleware } from './cached-nonce-tx-middleware'
 export { SpeculativeNonceTxMiddleware } from './speculative-nonce-tx-middleware'
 export { SignedEthTxMiddleware } from './signed-eth-tx-middleware'
-export { SignedTrxTxMiddleware } from './signed-trx-tx-middleware'
+export { SignedTrxTxMiddleware } from './signed-tron-tx-middleware'
